requirements.txt
quickfix==1.15.1
BitstampClient==2.2.8
